MDEV-24731 fixup: bogus assertion
DeadlockChecker::search(): Move a bogus assertion into a condition.
If the current transaction is waiting for a table lock (on something
else than an auto-increment lock), it is well possible that other
transactions are holding not only a conflicting lock, but also an
auto-increment lock.

This mistake was noticed during the testing of MDEV-24731, but it was
accidentally introduced in commit 5f463857645d8016e23afdb02955cdc305bebcfb.

lock_wait_end(): Remove an unused variable, and add an assertion.