MENT-1083: hashicorp plugin blocks server start
The plugin configuration file for hashicorp vault plugin
must be manually modified by the user in order to enter
the necessary parameters, such as the url to access the
vault server and the secret token. If these parameters
are not set properly, an attempt to connect to the server
is completed by timeout, which leads to a startup blocking.

This commit contains a modified configuration file that
prevents the plugin from loading until it is configured.
The user must manually configure all the necessary parameters
and explicitly uncomment the plugin loading by uncommenting
the corresponding line in the configuration file. Also
this fix is to change the maturity level to avoid an error
when loading the plugin (due to the fact that the maturity
level was now set to beta).