MDEV-24940: dict_hdr_get_new_id() changes table id in dictionary header, but does not write log record for it
Take a look dict_hdr_get_new_id(). If a table is temporary, log mode is
set to MTR_LOG_NO_REDO. But then table id is increased and written back
to space 0. But it's not logged due to log mode.

The proper fix is:
MDEV-17794 Do not assign persistent ID for temporary tables

But we don't want to backport a half of 10.[345] commits in order to fix
all bugs properly for this certain custom build. So this commit is just a
simple workaround for the bug fix.