MDEV-23497 fixup: Do not warn for ALTER TABLE conversion
Our intention is to make users aware that the ROW_FORMAT=COMPRESSED
format is on its way to deprecation.  It is an unnecessary annoyance
to users if we issue a warning when the user is converting a
ROW_FORMAT=COMPRESSED table to a supported format
(such as ROW_FORMAT=DYNAMIC).

ha_innobase::is_read_only(): Add a parameter to specify that an
ALTER TABLE to a supported format has been requested.

ha_innobase::check_if_supported_inplace_alter(): Set the parameter
when ALTER_OPTIONS to something else than ROW_FORMAT=COMPRESSED
(and without KEY_BLOCK_SIZE) is being requested.

Thanks to Elena Stepanova for suggesting this.