MDEV-23538: Rename mariadb.pc to mariadb-server-embedded.pc to avoid confusion
Change so that the client library config can be accessed using the name
mariadb.pc (or libmariadb.pc).

Rename old mariadb.pc to mariadb-embedded-server.pc to better indicate
that the config is for the daemon build, not client.

Also keep/include names with 'lib' as symlinks, as it aligns with how most
of *.pc files are named in general.

In downstream Debian we already ship the libmariadb.pc as mariadb.pc:
https://salsa.debian.org/mariadb-team/mariadb-10.5/-/commit/2f183af990fbe1cfa8c343998c7640f45f45368b

After this change we would have for the files:
- mariadb.pc for client (with libmariadb.pc as symlink)
- mariadb-embedded-sever.pc for server (with libmariadbd.pc as symlink)