MDEV-23634: Select query hanged the server and leads to OOM ...
Handle "col!=const" in the same way that MDEV-21958 did for
"col NOT IN(const-list)": do not consider it sargable if there
is a UNIQUE KEY(col).

The testcase is in main/range.test. The rest of test updates are
due to widespread use of 'pk!=1' in the testsuite. Changed the syntax
so that the conditions remain sargable.