MDEV-25328: --innodb command line option causes mariabackup to fail
This patch fixes an issue with launching mariabackup during SST
(when used with Galera), when during bootstrap mariabackup receives
the "--innodb" option, which is incorrectly interpreted as shortcut
for "--innodb-force-recovery". This patch does not require separate
test for mtr, as the problem is visible in general testing on
buildbot.

Also removed extra spaces in generated command lines (minor "cosmetic"
change to the patch that caused problem with "--innodb") and clarified
abbreviated option names in some tests, to avoid problems with ambiguous
options in the future.