MDEV-25425 Useless message "If the mysqld execution user is authorized page cleaner thread priority can be changed."
On startup, InnoDB tried to increase the priority of the page cleaner
thread. This would usually fail, resulting in a server log message.

When the page cleaner code was refactored in MDEV-23855, the throughput
and latency were greatly improved even when the
buf_flush_page_cleaner thread always ran the same priority with other
threads in our benchmarks.

So, this code should have no benefit in MariaDB Server 10.5.7 or later.

Besides, starting with MariaDB Server 10.5, the executable is called
mariadbd, not mysqld.